+++ Specs/Scene/ClassificationPrimitiveSpec.js
@@ -67,6 +67,7 @@ defineSuite([
         this._primitive = primitive;
         this.pass = Pass.GLOBE;
     }
+
     MockGlobePrimitive.prototype.update = function(frameState) {
         var commandList = frameState.commandList;
         var startLength = commandList.length;
@@ -996,4 +997,5 @@ defineSuite([
         primitive.destroy();
         expect(primitive.isDestroyed()).toEqual(true);
     });
+
 }, 'WebGL');